CDM-19 : moving CQL to its own package
