+++ src/main/java/datastax/astra/migrate/CopyJobSession.java
@@ -2,7 +2,7 @@ package datastax.astra.migrate;
 
 import com.datastax.oss.driver.api.core.CqlSession;
 import com.datastax.oss.driver.api.core.cql.*;
-import datastax.astra.migrate.properties.KnownProperties;
+import datastax.astra.migrate.cql.CqlHelper;
 import org.apache.spark.SparkConf;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
